TST: collect arithmetic test helpers
